move default authz provider classes to server module, and skip publishing of test-fixtures
