+++ buildSrc/src/main/groovy/com.flipkart.varadhi.java-published-library-conventions.gradle
@@ -4,7 +4,6 @@ plugins {
     id 'maven-publish'
 }
 
-
 publishing {
     publications {
         maven(MavenPublication) {
@@ -15,6 +14,8 @@ publishing {
 
             pom {
                 name = "${project.name}"
+                description = "Common AuthZ Provider interface for integrating with Varadhi"
+                url = "https://github.com/flipkart-incubator/varadhi"
             }
         }
     }
