+++ core/src/main/java/com/flipkart/varadhi/spi/db/MetaStore.java
@@ -1,5 +1,6 @@
 package com.flipkart.varadhi.spi.db;
 
+import com.flipkart.varadhi.auth.RoleBindingNode;
 import com.flipkart.varadhi.entities.*;
 
 import java.util.List;
