+++ gradle.properties
@@ -3,5 +3,5 @@ projectName=varadhi
 projectVersion=0
 majorVersion=1
 minorVersion=0
-patchId=2-SNAPSHOT
+patchId=3-SNAPSHOT
 mavenPublishUrl="" # Add repository URL with double quotes removed
