+++ server/src/main/java/com/flipkart/varadhi/auth/DefaultAuthorizationProvider.java
@@ -1,5 +1,6 @@
-package com.flipkart.varadhi.authz;
+package com.flipkart.varadhi.auth;
 
+import com.flipkart.varadhi.authz.AuthorizationProvider;
 import com.flipkart.varadhi.config.AuthorizationOptions;
 import com.flipkart.varadhi.config.DefaultAuthorizationConfiguration;
 import com.flipkart.varadhi.entities.*;
