+++ server/src/main/java/com/flipkart/varadhi/entities/Role.java
