+++ server/src/main/resources/authorizationConfig.yml
