+++ server/src/test/java/com/flipkart/varadhi/auth/DefaultAuthorizationProviderTest.java
@@ -1,9 +1,9 @@
-package com.flipkart.varadhi.authz;
+package com.flipkart.varadhi.auth;
 
 import com.flipkart.varadhi.config.AuthorizationOptions;
+import com.flipkart.varadhi.entities.ResourceAction;
 import com.flipkart.varadhi.exceptions.InvalidConfigException;
 import io.vertx.core.Vertx;
-import com.flipkart.varadhi.entities.ResourceAction;
 import io.vertx.junit5.Checkpoint;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
